pyproject.toml
[project]
name = "sts-judge-api"
version = "0.0.1"
description = "STS API"
authors = [{ name = "DoumanAsh", email = "douman@gmx.se" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
dependencies = [
    "django>=4.2.24",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.16.1",
    "huggingface-hub[hf-xet]>=0.34.4",
    "langdetect>=1.0.9",
    "pytest-mock>=3.15.0",
    "sentence-transformers>=5.1.0",
]

[project.urls]
Homepage = "https://DoumanAsh.github.io/sts-judge-api/"
Repository = "https://github.com/DoumanAsh/sts-judge-api"
Documentation = "https://DoumanAsh.github.io/sts-judge-api/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "pyright>=1.1.405",
    "pytest-django>=4.11.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api/", "internal/"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.9"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "api.settings"
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["api"]
